requirements.txt
cohere
faiss-cpu==1.7.4
openai
sentence-transformers
numpy
pandas
fastapi
Uvicorn[all]
